# backend-express/swagger.yaml
openapi: 3.0.0
info:
  title: Kopi Haus Dashboard API
  version: 1.0.0
  description: API documentation for the Kopi Haus Dashboard backend application.
servers:
  - url: http://localhost:5001/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Server error
...
    Product:
      type: object
      required:
        - sku
        - name
        - price
        - cost
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c6
        sku:
          type: string
          example: P001
        name:
          type: string
          example: Espresso
        category_id:
          type: string
          nullable: true
          example: 60d5ec49f8c7a40015a7b8c2
        unit:
          type: string
          example: cup
        price:
          type: number
          format: float
          example: 25000
        cost:
          type: number
          format: float
          example: 10000
        tax_id:
          type: string
          nullable: true
          example: 60d5ec49f8c7a40015a7b8c5
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T10:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T10:00:00.000Z'
...
    ProductPrice:
      type: object
      required:
        - product_id
        - outlet_id
        - price
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c8
        product_id:
          type: string
          example: 60d5ec49f8c7a40015a7b8c6
        outlet_id:
          type: string
          example: 60d5ec49f8c7a40015a7b8c1
        price:
          type: number
          format: float
          example: 27500
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T10:00:00.000Z'
...
paths:
...
  /products:
    get:
      summary: Get all products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sku
                - name
                - price
                - cost
              properties:
                sku:
                  type: string
                  example: P002
                name:
                  type: string
                  example: Latte
                category_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c2
                unit:
                  type: string
                  example: cup
                price:
                  type: number
                  format: float
                  example: 30000
                cost:
                  type: number
                  format: float
                  example: 12000
                tax_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c5
                is_active:
                  type: boolean
                  example: true
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input, category ID, or tax ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      summary: Get a product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the product to retrieve
      responses:
        '200':
          description: Product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update a product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the product to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sku:
                  type: string
                  example: P002-UPD
                name:
                  type: string
                  example: Updated Latte
                category_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c2
                unit:
                  type: string
                  example: cup
                price:
                  type: number
                  format: float
                  example: 32000
                cost:
                  type: number
                  format: float
                  example: 13000
                tax_id:
                  type: string
                  nullable: true
                  example: 60d5ec49f8c7a40015a7b8c5
                is_active:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input, category ID, or tax ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the product to delete
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /product-prices:
    get:
      summary: Get all product prices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of product prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductPrice'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new product price
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - outlet_id
                - price
              properties:
                product_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c6
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                price:
                  type: number
                  format: float
                  example: 27500
      responses:
        '201':
          description: Product price created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPrice'
        '400':
          description: Invalid input, product ID, or outlet ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /product-prices/{id}:
    get:
      summary: Get a product price by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the product price to retrieve
      responses:
        '200':
          description: Product price data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPrice'
        '404':
          description: Product price not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update a product price by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the product price to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c6
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                price:
                  type: number
                  format: float
                  example: 28000
      responses:
        '200':
          description: Product price updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPrice'
        '400':
          description: Invalid input, product ID, or outlet ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product price not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a product price by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the product price to delete
      responses:
        '200':
          description: Product price deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Product price not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...